requirements.txt
kagglehub
numpy
pandas
opencv-python
pillow
matplotlib
scikit-learn
torch
torchvision
tqdm
albumentations